+++ src/meta/src/raft/node.rs
@@ -18,8 +18,8 @@ pub enum NodeState {
 #[derive(Clone)]
 pub struct Node {
     pub node_ip: String,
-    pub node_id: i32,
-    pub leader_id: Option<i32>,
+    pub node_id: u64,
+    pub leader_id: Option<u64>,
     pub leader_ip: Option<String>,
     pub raft_state: NodeRaftState,
     pub state: NodeState,
@@ -27,7 +27,7 @@ pub struct Node {
 }
 
 impl Node {
-    pub fn new(node_ip: String, node_id: i32) -> Node {
+    pub fn new(node_ip: String, node_id: u64) -> Node {
         Node {
             node_ip,
             node_id,
